put shared utils in vis folder
